bugfix inventory item structs and items on belts
