+++ src/index.ts
@@ -2,6 +2,7 @@
  * @file Automatically generated by barrelsby.
  */
 
+export * from "./parser/parser";
 export * from "./parser/byte/alignment.enum";
 export * from "./parser/byte/binary-operable.interface";
 export * from "./parser/byte/binary-readable.interface";
@@ -13,22 +14,21 @@ export * from "./parser/context/context-writer";
 export * from "./parser/context/hierarchical-version-context";
 export * from "./parser/context/reader-writer-context";
 export * from "./parser/error/parser.error";
-export * from "./parser/parser";
-export * from "./parser/satisfactory/blueprint/blueprint-config";
 export * from "./parser/satisfactory/blueprint/blueprint-config-version";
-export * from "./parser/satisfactory/blueprint/blueprint-header";
+export * from "./parser/satisfactory/blueprint/blueprint-config";
 export * from "./parser/satisfactory/blueprint/blueprint-header-version";
+export * from "./parser/satisfactory/blueprint/blueprint-header";
 export * from "./parser/satisfactory/blueprint/blueprint-reader";
 export * from "./parser/satisfactory/blueprint/blueprint-writer";
 export * from "./parser/satisfactory/blueprint/blueprint.types";
 export * from "./parser/satisfactory/edit/edit-constants";
 export * from "./parser/satisfactory/save/data-blob";
 export * from "./parser/satisfactory/save/factory-game-custom-version";
-export * from "./parser/satisfactory/save/level";
 export * from "./parser/satisfactory/save/level-to-destroyed-actors-map";
+export * from "./parser/satisfactory/save/level";
 export * from "./parser/satisfactory/save/object-references-list";
-export * from "./parser/satisfactory/save/satisfactory-save";
 export * from "./parser/satisfactory/save/satisfactory-save-header";
+export * from "./parser/satisfactory/save/satisfactory-save";
 export * from "./parser/satisfactory/save/save-body-chunks";
 export * from "./parser/satisfactory/save/save-custom-version";
 export * from "./parser/satisfactory/save/save-header-type";
@@ -39,15 +39,10 @@ export * from "./parser/satisfactory/save/toc-blob";
 export * from "./parser/satisfactory/types/objects/SaveComponent";
 export * from "./parser/satisfactory/types/objects/SaveEntity";
 export * from "./parser/satisfactory/types/objects/SaveObject";
+export * from "./parser/satisfactory/types/property/PropertiesList";
 export * from "./parser/satisfactory/types/property/generic/AbstractBaseProperty";
 export * from "./parser/satisfactory/types/property/generic/BoolProperty";
 export * from "./parser/satisfactory/types/property/generic/ByteProperty";
-export * from "./parser/satisfactory/types/property/generic/containers/ArrayProperty";
-export * from "./parser/satisfactory/types/property/generic/containers/ArrayPropertySpecificTypes";
-export * from "./parser/satisfactory/types/property/generic/containers/MapProperty";
-export * from "./parser/satisfactory/types/property/generic/containers/SetProperty";
-export * from "./parser/satisfactory/types/property/generic/containers/SetPropertySpecificTypes";
-export * from "./parser/satisfactory/types/property/generic/containers/StructProperty";
 export * from "./parser/satisfactory/types/property/generic/DoubleProperty";
 export * from "./parser/satisfactory/types/property/generic/EnumProperty";
 export * from "./parser/satisfactory/types/property/generic/FloatProperty";
@@ -64,7 +59,12 @@ export * from "./parser/satisfactory/types/property/generic/TextProperty";
 export * from "./parser/satisfactory/types/property/generic/Uint32Property";
 export * from "./parser/satisfactory/types/property/generic/Uint64Property";
 export * from "./parser/satisfactory/types/property/generic/Uint8Property";
-export * from "./parser/satisfactory/types/property/PropertiesList";
+export * from "./parser/satisfactory/types/property/generic/containers/ArrayProperty";
+export * from "./parser/satisfactory/types/property/generic/containers/ArrayPropertySpecificTypes";
+export * from "./parser/satisfactory/types/property/generic/containers/MapProperty";
+export * from "./parser/satisfactory/types/property/generic/containers/SetProperty";
+export * from "./parser/satisfactory/types/property/generic/containers/SetPropertySpecificTypes";
+export * from "./parser/satisfactory/types/property/generic/containers/StructProperty";
 export * from "./parser/satisfactory/types/property/special/BuildableSubsystemSpecialProperties";
 export * from "./parser/satisfactory/types/property/special/CircuitSpecialProperties";
 export * from "./parser/satisfactory/types/property/special/ConveyorChainActorSpecialProperties";
@@ -77,11 +77,27 @@ export * from "./parser/satisfactory/types/property/special/runtime-buildable-in
 export * from "./parser/satisfactory/types/property/special/SpecialDroneActionProperties";
 export * from "./parser/satisfactory/types/property/special/SpecialProperties";
 export * from "./parser/satisfactory/types/property/special/VehicleSpecialProperties";
+export * from "./parser/satisfactory/types/structs/col4";
+export * from "./parser/satisfactory/types/structs/DynamicStructPropertyValue";
+export * from "./parser/satisfactory/types/structs/FGDynamicStruct";
+export * from "./parser/satisfactory/types/structs/GUIDInfo";
+export * from "./parser/satisfactory/types/structs/MD5Hash";
+export * from "./parser/satisfactory/types/structs/ObjectReference";
+export * from "./parser/satisfactory/types/structs/SaveBodyValidation";
+export * from "./parser/satisfactory/types/structs/SoftObjectReference";
+export * from "./parser/satisfactory/types/structs/Transform";
+export * from "./parser/satisfactory/types/structs/vec2";
+export * from "./parser/satisfactory/types/structs/vec3";
+export * from "./parser/satisfactory/types/structs/vec4";
+export * from "./parser/satisfactory/types/structs/VehiclePhysicsData";
 export * from "./parser/satisfactory/types/structs/binary/FClientIdentityInfo";
 export * from "./parser/satisfactory/types/structs/binary/FColor";
+export * from "./parser/satisfactory/types/structs/binary/FConveyorBeltItem";
+export * from "./parser/satisfactory/types/structs/binary/FConveyorChainSplineSegment";
 export * from "./parser/satisfactory/types/structs/binary/FCustomVersion";
 export * from "./parser/satisfactory/types/structs/binary/FCustomVersionContainer";
 export * from "./parser/satisfactory/types/structs/binary/FEngineVersion";
+export * from "./parser/satisfactory/types/structs/binary/FInventoryItem";
 export * from "./parser/satisfactory/types/structs/binary/FLinearColor";
 export * from "./parser/satisfactory/types/structs/binary/FLocalUserNetIdBundle";
 export * from "./parser/satisfactory/types/structs/binary/FPackageFileVersion";
@@ -92,26 +108,13 @@ export * from "./parser/satisfactory/types/structs/binary/FSaveObjectVersionData
 export * from "./parser/satisfactory/types/structs/binary/FUniqueNetIdRepl";
 export * from "./parser/satisfactory/types/structs/binary/GUID";
 export * from "./parser/satisfactory/types/structs/binary/LBBalancerIndexing";
-export * from "./parser/satisfactory/types/structs/col4";
-export * from "./parser/satisfactory/types/structs/DynamicStructPropertyValue";
-export * from "./parser/satisfactory/types/structs/FGDynamicStruct";
-export * from "./parser/satisfactory/types/structs/GUIDInfo";
-export * from "./parser/satisfactory/types/structs/MD5Hash";
 export * from "./parser/satisfactory/types/structs/mods/FicsItCam/FICFrameRange";
 export * from "./parser/satisfactory/types/structs/mods/FicsItNetworks/FINGPUT1BufferPixel";
 export * from "./parser/satisfactory/types/structs/mods/FicsItNetworks/FINItemStateFileSystem";
 export * from "./parser/satisfactory/types/structs/mods/FicsItNetworks/FINLuaRuntimePersistenceState";
 export * from "./parser/satisfactory/types/structs/mods/FicsItNetworks/FINNetworkTrace";
 export * from "./parser/satisfactory/types/structs/mods/ModularLoadBalancers/FLBBalancerIndexing";
-export * from "./parser/satisfactory/types/structs/ObjectReference";
-export * from "./parser/satisfactory/types/structs/SaveBodyValidation";
-export * from "./parser/satisfactory/types/structs/SoftObjectReference";
-export * from "./parser/satisfactory/types/structs/Transform";
-export * from "./parser/satisfactory/types/structs/vec2";
-export * from "./parser/satisfactory/types/structs/vec3";
-export * from "./parser/satisfactory/types/structs/vec4";
-export * from "./parser/satisfactory/types/structs/VehiclePhysicsData";
 export * from "./parser/stream/reworked/readable-stream-parser";
 export * from "./parser/unreal-engine/ETextHistoryType";
 export * from "./parser/unreal-engine/EUnrealEngineObjectUE5Version";
-
+export * from "./test/parser.spec";
